build: build ceph dependencies from source
With this commit we now build some of the packages ceph depends
on directly from source code. We use cmake and externalproject_add
to build the sources using a superbuild project approach. Not all
packages that are compiled into rook are built from source yet, but
enough are there to work around some of the issues we were ran into
with adding RGW to rook.

Specificaly, libcurl is now built from source and removes the
dependency on openldap and other unused protocols. Boost, zlib,
libblkid, and libuuid are now also built from source. More packages
will be added in the future to complete this approach.

The build container is now based on the yakkety release of ubuntu.