[doc] Remedy falsehood about losing up to 2/3 of a cluster
First, this depends on the MIN_SIZE parameter of a pool. Second even if the
MIN_SIZE is one this would only hold up to six node cluster. Imagine a nine
node cluster; losing two-thirds of a cluster can lead to a situation where all
 replicas of an object are offline. I.e., you can only lose up to two failure
domains, in this case nodes, when you have a replication size of three.