Ensuring volume lock is released when the same pod is requesting it
There is a case where a pod is trying to request a volume
attachment after the previous attachment failed. However even though this is
the same pod as the original attachment request, its UID could be different. This is normally caused when
the user manually deletes the pod and recreate it again. The name will be the same but the ID will be different

fixes https://github.com/rook/rook/issues/1214

(cherry picked from commit c1882d4760f5148d7c66dd4303e0990d5f6bfcfd)