Give ceph more exposure
This is an attempt to avoid confusion for the end users.
Rook is not the storage entity, it's the orchestrator that currently deploys Ceph.
Other backends might follow but now everything (deployment, code,
examples) is being done with Ceph.

Ceph is the storage technology so anything related to configuration
should highlight Ceph and not Rook.

I believe it's crucial that users make the distinction between the
rook-operator/agent AND the Ceph cluster (mon,osd,rgw,mds,mgr etc).

This commits helps clarifying this by naming rook all the
variables/options related to the deployment and 'ceph' what's related to
the storage.

Signed-off-by: Sébastien Han <seb@redhat.com>