Properly terminate controller threads on cluster delete
The stopCh value was only present in struct cluster, whose instance was
local to the onAdd function; save it to ClusterController.

Second, it was not initialized.

Third, ClusterController's "devicesInUse" needs to be set false on
cluster delete.

With this fix, Ceph clusters can be deleted and recreated without
restarting the Rook operator.

Signed-off-by: Dan Mick <dan.mick@redhat.com>