Image make: Move save image step to release builds
The image save step in the image build makefiles takes half or more of
the time of builds, and this step is not necessary for development
builds. Move this functionality to the `build/release/Makefile` so that
images are saved during release builds but not during dev builds. This
will allow developers to iterate more quickly on their builds.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>