nfs: rebase nfs image onto fedora 28
f26 has been EOL'ed for a while now, so we shouldn't be relying on it.

Update to fedora 28. This necessitates installing libnsl2 for ganesha,
which is not installed in the base image these days.

Long term, we should probably find some way to keep this in sync with
the Fedora release cadence, or rebase onto something more stable like
Centos.

Signed-off-by: Jeff Layton <jlayton@redhat.com>