Fix the test function that deletes the replicapool
The command `kubectl delete -n rook-ceph pool replicapool` used to
work, but currently returns the error `error: the server doesn't
have a resource type "pool"`. The actual resource type is
`cephblockpools`. The resource type `cluster` was also recently
changed to `cephcluster`.

Since the function commands end with `|| true`, the errors are
silently masked and the resources are not deleted.

This PR adds delete commands for `cephblockpools` and `cephcluster`,
so that regardless of whether you're testing old or new versions
of Rook the delete commands will work.

Signed-off-by: Earl C. Ruby III <earlruby@users.noreply.github.com>