Document the RBD module prerequisite for Rook Ceph
The RBD module is an important prerequisite for Rook Ceph. It is mentioned as a possible problem on the Common Issues page, but that is the ambulance at the bottom of the cliff 😄. We should mention the perquisite for new users, tell them how to check, and note any major Kubernetes-as-a-service providers that do not have the RBD module.
https://rook.io/docs/rook/master/common-issues.html#rook-agent-rbd-module-missing-error

Signed-off-by: Aaron Roydhouse <aaron@roydhouse.com>