ceph mon: revise start/update procedures
Revise the start/update procedures for Ceph mons to be safer.

When starting new mons, the mons are now created all at once, and then
the mon operator module will report success once all mons have achieved
quorum.

When the operator starts and there are existing mons, the operator first
waits for all of the existing mons to be in quorum. After they are in
quorum, the update procedure commences, where each mon is updated if the
new deployment differs from the existing deployment. If a deployment is
updated, the update will pause until all mons are again in quorum before
continuing to update the next mon.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>