ceph osd: ceph-vol activate in init container
Make some small changes to the Ceph osd daemon pod that runs daemons for
ceph-volume-provisioned osds to run the `ceph-volume lvm activate...`
command and `ceph-osd...` command directly from the container
entrypoints. This is one small step in transitioning the OSDs to be
configured as much as possible from the operator.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>