cassandra: replace_address_first_boot safety measure for data loss
To protect against data loss that can result from the failure of a local disk
or Node, we adopt the following procedure:

Each time a Member starts, the sidecar will:
- Check if it bootstrapped by contacting another Member of the Cluster.
  - If it isn't bootstrapped, it will start normally.
  - If it is bootstrapped, it will be started with the "replace_address_first_boot"
    option.

The "replace_address_first_boot" option will check if the Member has lost its data.
If it has, it will assume the tokens of the fallen Member and its data.
Otherwise, the Member will start normally.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>