Don't automatically use filestore on systems with all devices absent.
On setups where all specified disks are absent, we shouldn't
automatically make use of filestore. It may well be the case that the
disks still need to be installed.

There is already code in place to realise this, except that it checks
the list of OSDs that end up being configured, as opposed to the ones
that are specified in configuration.

Fixes: #1007
Signed-off-by: Ed Schouten <ed.schouten@prodrive-technologies.com>