script: add rook-ceph-dev script
That's the script I use to develop on Rook. I'm sharing it in the hope
that it'd be useful to people.

This script assumes that kubernetes is up and running and works only
with minikube as well as using YAMLs from the .idea directory.

Oh and it only works for Ceph.

Signed-off-by: Sébastien Han <seb@redhat.com>