build: publish the master build before integration tests
The master build will be published as soon as the basic build and unit tests
are completed. This ensures that the generated help documentation matches
what is in master and the latest master published build matches
the yaml files in master. The longer the delay between merging and
publishing, the more likely users are to run into issues with
test deployments. This change minimizes the gap.

Signed-off-by: travisn <tnielsen@redhat.com>