add mergify engine
The mergify engine will help us merging PRs without special permissions
on the repo. Being part of the organization is enough. So if someone
from the orgs approuves a PR, it can be merged. That person does not
need to be a maintainer and/or admin of the github org.

The goal is to make smoother progress on merging PRs without being
blocked waiting for a maintainer to merge it.

This basically:

* allow a smoother merge experience
* remove the maintainer bottleneck is the person is away
* release pressure off the maintainers

We can go deeper with who is taken into account for a merge, see https://doc.mergify.io/conditions.html#attribute-list

Signed-off-by: Sébastien Han <seb@redhat.com>