ceph: selectively apply http bind addr fix
fixes: #2492

the http bind address fix in rook address an upstream issue in ceph in
which cherrypy break when binding to `::` when the host doesn't have an
ipv6 address. it fixes the issue by always binding prometheus and the
dashboard webservers to the pod address. by doing this, it prevents
using the kubectl port-forward technique, which some users would like.

a patch to cherrypy is upstream in ceph. this PR checks the ceph version
and selectively disables the fix so that both port-forward and the
normal service method works. this is the verison matrix:

          FIX     VER
          ---     ---
luminous merged (>= 12.2.12)
mimic    merged (>= 13.2.6)
nautilus merged (>= 14.1.1)

when all the containers for these versions exists and sufficient time
passes the unreachable code guarded by the version checks in this patch
can be removed.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>