ceph: Handle cluster crates/updates in a goroutine
The create and update event handlers for the CephCluster Controller will
launch a new goroutine for handling each new event. If an previously
existing goroutine is found for a cluster, it will stop it before
launching a new goroutine.

With this, the update event handler will no longer be blocked by long
timeouts (1h) when handling faulty CR updates.

Fixes #3047.

Signed-off-by: Kaushal M <kshlmster@gmail.com>